create new package steep_icecavity and modify verif experiments
