+++ pkg/pkg_depend
@@ -43,4 +43,5 @@ ecco		+mdsio		+cal
 tapenade	+autodiff
 grdchk		+autodiff	+cost		+ctrl
 showflops	+runclock
+steep_cavity    +shelfice
