+++ pkg/shelfice/SHELFICE.h
@@ -49,7 +49,6 @@ C     SHELFICEsaltTransCoeff   :: constant salinity transfer coefficient that
 C                                 determines salt flux into shelfice
 C                                 (def: SHELFICEsaltToHeatRatio * SHELFICEheatTransCoeff)
 C     -----------------------------------------------------------------------
-C     SHELFICEuseSTIC          :: use steep ice cavity code (STIC)
 C     SHELFICEuseGammaFrict    :: use velocity dependent exchange coefficients,
 C                                 see Holland and Jenkins (1999), eq.11-18,
 C                                 with the following parameters (def: F):
@@ -168,12 +167,12 @@ CEOP
       _RL shelficeLoadAnomaly   (1-OLx:sNx+OLx,1-OLy:sNy+OLy,nSx,nSy)
       _RL shelficeForcingT      (1-OLx:sNx+OLx,1-OLy:sNy+OLy,nSx,nSy)
       _RL shelficeForcingS      (1-OLx:sNx+OLx,1-OLy:sNy+OLy,nSx,nSy)
-#ifndef ALLOW_SHITRANSCOEFF_3D
-      _RL shiTransCoeffT        (1-OLx:sNx+OLx,1-OLy:sNy+OLy,nSx,nSy)
-      _RL shiTransCoeffS        (1-OLx:sNx+OLx,1-OLy:sNy+OLy,nSx,nSy)
-#else
+#ifdef ALLOW_SHITRANSCOEFF_3D
       _RL shiTransCoeffT     (1-OLx:sNx+OLx,1-OLy:sNy+OLy,Nr,nSx,nSy)
       _RL shiTransCoeffS     (1-OLx:sNx+OLx,1-OLy:sNy+OLy,Nr,nSx,nSy)
+#else
+      _RL shiTransCoeffT        (1-OLx:sNx+OLx,1-OLy:sNy+OLy,nSx,nSy)
+      _RL shiTransCoeffS        (1-OLx:sNx+OLx,1-OLy:sNy+OLy,nSx,nSy)
 #endif
       _RL shiCDragFld        (1-OLx:sNx+OLx,1-OLy:sNy+OLy,nSx,nSy)
       _RL shiDragQuadFld     (1-OLx:sNx+OLx,1-OLy:sNy+OLy,nSx,nSy)
@@ -201,7 +200,6 @@ CEOP
 #endif /* ALLOW_DIAGNOSTICS */
 
       LOGICAL SHELFICEisOn
-      LOGICAL SHELFICEuseSTIC
       LOGICAL useISOMIPTD
       LOGICAL SHELFICEconserve
       LOGICAL SHELFICEboundaryLayer
@@ -221,7 +219,6 @@ CEOP
       COMMON /SHELFICE_PARMS_L/
      &     SHELFICEisOn,
      &     useISOMIPTD,
-     &     SHELFICEuseSTIC,
      &     SHELFICEconserve,
      &     SHELFICEboundaryLayer,
      &     SHI_withBL_realFWflux,
@@ -250,79 +247,4 @@ CEOP
      &     SHELFICEMassDynTendFile,
      &     SHELFICETransCoeffTFile
 
-#ifdef ALLOW_STEEP_ICECAVITY
-C--   Constants that can be set in data.shelfice
-C     icfThetaHorizDiffusionLength ::
-C     icfThetaInterior  ::
-C     STICdepthFile     :: name of icefront depth file (m)
-C                          2D file containing depth of the ice front
-C                          at each model grid cell
-C     STIClengthFile    :: name of icefront length file (m/m^2)
-C                          2D file containing the ratio of the horizontal
-C                          length of the ice front in each model grid cell
-C                          divided by the grid cell area
-
-      COMMON /STIC_PARMS_R/
-     &     icfThetaHorizDiffusionLength,
-     &     icfThetaInterior
-      _RL icfThetaHorizDiffusionLength
-      _RL icfThetaInterior
-
-      CHARACTER*(MAX_LEN_FNAM) STIClengthFile
-      CHARACTER*(MAX_LEN_FNAM) STICdepthFile
-      COMMON /STIC_PARM_C/
-     &     STIClengthFile,
-     &     STICdepthFile
-
-C     ow - 06/29/2018
-C     maskSHI of pkg/shelflice is not consistent with the spirit of gencost.
-C       Use the following masks below instead.
-C     mask2dSHIICF: 2d shelf-ice & ice-front mask:
-C       1 for having shelf-ice and/or ice-front at one or more vertical levels
-C       and 0 otherwise.
-C     mask3dSHIICF: 3d shelf-ice & ice-front mask.
-C     mask2dSHI: 2d shelf-ice mask
-C     mask3dSHI: 3d shelf-ice mask
-C     mask2dICF: 2d ice-front mask: 1 for having ice-front at one or more
-C                vertical levels.
-C     mask3dICF: 3d ice-front mask
-      COMMON /STIC_MASKS/ mask2dSHIICF, mask3dSHIICF,
-     &        mask2dSHI, mask3dSHI, mask2dICF, mask3dICF
-      _RS mask2dSHIICF  (1-OLx:sNx+OLx,1-OLy:sNy+OLy,nSx,nSy)
-      _RS mask3dSHIICF  (1-OLx:sNx+OLx,1-OLy:sNy+OLy,Nr,nSx,nSy)
-      _RS mask2dSHI  (1-OLx:sNx+OLx,1-OLy:sNy+OLy,nSx,nSy)
-      _RS mask3dSHI  (1-OLx:sNx+OLx,1-OLy:sNy+OLy,Nr,nSx,nSy)
-      _RS mask2dICF  (1-OLx:sNx+OLx,1-OLy:sNy+OLy,nSx,nSy)
-      _RS mask3dICF  (1-OLx:sNx+OLx,1-OLy:sNy+OLy,Nr,nSx,nSy)
-
-C     INTEGER arrays
-C     CURI_ARR       :: i-index for neighboring ice-front points
-C     CURJ_ARR       :: j-index for neighboring ice-front points
-C     sticfWidth_arr :: ice-front width in meters
-C     kIcf           :: index of the bottommost ice front cell (2D)
-      COMMON /STIC_ICEFRONT_I/ kIcf, CURI_ARR, CURJ_ARR
-      INTEGER kIcf    (1-OLx:sNx+OLx,1-OLy:sNy+OLy,nSx,nSy)
-      INTEGER CURI_ARR(1-OLx:sNx+OLx,1-OLy:sNy+OLy,nSx,nSy,4)
-      INTEGER CURJ_ARR(1-OLx:sNx+OLx,1-OLy:sNy+OLy,nSx,nSy,4)
-
-      COMMON /STIC_FIELDS_RL/
-     &     icfForcingT, icfForcingS,
-     &     sticfWidth_arr
-      _RL icfForcingT   (1-OLx:sNx+OLx,1-OLy:sNy+OLy,Nr,nSx,nSy)
-      _RL icfForcingS   (1-OLx:sNx+OLx,1-OLy:sNy+OLy,Nr,nSx,nSy)
-      _RL sticfWidth_arr(1-OLx:sNx+OLx,1-OLy:sNy+OLy,nSx,nSy,4)
-
-
-      COMMON /STIC_FIELDS_RS/
-     &     R_stic, sticfLength,
-     &     icfHeatFlux, icfFreshWaterFlux,
-     &     sticfHeatFlux, sticfFreshWaterFlux
-      _RS R_stic             (1-OLx:sNx+OLx,1-OLy:sNy+OLy,nSx,nSy)
-      _RS sticfLength        (1-OLx:sNx+OLx,1-OLy:sNy+OLy,nSx,nSy)
-      _RL icfHeatFlux        (1-OLx:sNx+OLx,1-OLy:sNy+OLy,Nr,nSx,nSy)
-      _RL icfFreshWaterFlux  (1-OLx:sNx+OLx,1-OLy:sNy+OLy,Nr,nSx,nSy)
-      _RL sticfHeatFlux      (1-OLx:sNx+OLx,1-OLy:sNy+OLy,nSx,nSy)
-      _RL sticfFreshWaterFlux(1-OLx:sNx+OLx,1-OLy:sNy+OLy,nSx,nSy)
-
-#endif /* ALLOW_STEEP_ICECAVITY */
 #endif /* ALLOW_SHELFICE */
