+++ pkg/shelfice/shelfice_init_varia.F
@@ -39,29 +39,18 @@ CEOP
            shelficeForcingS      (i,j,bi,bj) = 0. _d 0
            shelficeHeatFlux      (i,j,bi,bj) = 0. _d 0
            shelficeFreshWaterFlux(i,j,bi,bj) = 0. _d 0
-#ifdef ALLOW_STEEP_ICECAVITY
-           sticfHeatFlux         (i,j,bi,bj) = 0. _d 0
-           sticfFreshWaterFlux   (i,j,bi,bj) = 0. _d 0
-#endif
-#ifndef ALLOW_SHITRANSCOEFF_3D
-           shiTransCoeffT        (i,j,bi,bj) = 0. _d 0
-           shiTransCoeffS        (i,j,bi,bj) = 0. _d 0
-#endif
            shelficeMass          (i,j,bi,bj) = 0. _d 0
            shiCDragFld           (i,j,bi,bj) = 0. _d 0
            shiDragQuadFld        (i,j,bi,bj) = 0. _d 0
-           DO k = 1, NR
 #ifdef ALLOW_SHITRANSCOEFF_3D
+           DO k = 1, NR
             shiTransCoeffT     (i,j,k,bi,bj) = 0. _d 0
             shiTransCoeffS     (i,j,k,bi,bj) = 0. _d 0
-#endif
-#ifdef ALLOW_STEEP_ICECAVITY
-            icfHeatFlux        (i,j,k,bi,bj) = 0. _d 0
-            icfFreshWaterFlux  (i,j,k,bi,bj) = 0. _d 0
-            icfForcingT        (i,j,k,bi,bj) = 0. _d 0
-            icfForcingS        (i,j,k,bi,bj) = 0. _d 0
-#endif
            ENDDO
+#else
+           shiTransCoeffT        (i,j,bi,bj) = 0. _d 0
+           shiTransCoeffS        (i,j,bi,bj) = 0. _d 0
+#endif
          ENDDO
         ENDDO
        ENDDO
@@ -76,27 +65,27 @@ C        Therefore, in this case, the "If useShelfIce" is added back here:
 
       IF ( .NOT. SHELFICEuseGammaFrict ) THEN
        IF ( SHELFICETransCoeffTFile .NE. ' ' ) THEN
-#ifndef ALLOW_SHITRANSCOEFF_3D
-        CALL READ_FLD_XY_RL( SHELFICETransCoeffTFile, ' ',
-     &                       shiTransCoeffT, 0, myThid )
-        _EXCH_XY_RL( shiTransCoeffT, myThid )
-#else
+#ifdef ALLOW_SHITRANSCOEFF_3D
         CALL READ_FLD_XYZ_RL( SHELFICETransCoeffTFile, ' ',
      &                       shiTransCoeffT, 0, myThid )
         _EXCH_XYZ_RL( shiTransCoeffT, myThid )
+#else
+        CALL READ_FLD_XY_RL( SHELFICETransCoeffTFile, ' ',
+     &                       shiTransCoeffT, 0, myThid )
+        _EXCH_XY_RL( shiTransCoeffT, myThid )
 #endif
         DO bj = myByLo(myThid), myByHi(myThid)
          DO bi = myBxLo(myThid), myBxHi(myThid)
           DO j = 1-OLy, sNy+OLy
            DO i = 1-OLx, sNx+OLx
-#ifndef ALLOW_SHITRANSCOEFF_3D
-            shiTransCoeffS(i,j,bi,bj) = SHELFICEsaltToHeatRatio *
-     &                                  shiTransCoeffT(i,j,bi,bj)
-#else
+#ifdef ALLOW_SHITRANSCOEFF_3D
             DO k = 1, Nr
             shiTransCoeffS(i,j,k,bi,bj) = SHELFICEsaltToHeatRatio *
      &                                  shiTransCoeffT(i,j,k,bi,bj)
             ENDDO
+#else
+            shiTransCoeffS(i,j,bi,bj) = SHELFICEsaltToHeatRatio *
+     &                                  shiTransCoeffT(i,j,bi,bj)
 #endif
            ENDDO
           ENDDO
@@ -108,14 +97,14 @@ C     set coefficients to constant values
          DO bi = myBxLo(myThid), myBxHi(myThid)
           DO j = 1-OLy, sNy+OLy
            DO i = 1-OLx, sNx+OLx
-#ifndef ALLOW_SHITRANSCOEFF_3D
-            shiTransCoeffT(i,j,bi,bj) = SHELFICEheatTransCoeff
-            shiTransCoeffS(i,j,bi,bj) = SHELFICEsaltTransCoeff
-#else
+#ifdef ALLOW_SHITRANSCOEFF_3D
             DO k = 1, Nr
              shiTransCoeffT(i,j,k,bi,bj) = SHELFICEheatTransCoeff
              shiTransCoeffS(i,j,k,bi,bj) = SHELFICEsaltTransCoeff
             ENDDO
+#else
+            shiTransCoeffT(i,j,bi,bj) = SHELFICEheatTransCoeff
+            shiTransCoeffS(i,j,bi,bj) = SHELFICEsaltTransCoeff
 #endif
            ENDDO
           ENDDO
