+++ pkg/steep_icecavity/stic_ad_diff.list
@@ -1,0 +1,3 @@
+stic_init_varia.f
+stic_thermodynamics.f
+stic_solve4fluxes.f
