+++ verification/isomip/code/packages.conf
@@ -4,6 +4,7 @@ ggl90
 cd_code
 obcs
 shelfice
+steep_icecavity
 icefront
 diagnostics
 mnc
