+++ verification/isomip/code_ad/SHELFICE_OPTIONS.h
@@ -14,8 +14,8 @@ C     *==========================================================*
 #ifdef ALLOW_SHELFICE
 C     Package-specific Options & Macros go here
 
-#define ALLOW_STEEP_ICECAVITY
-c#define ALLOW_SHITRANSCOEFF_3D
+C     New flag to be documented
+#undef ALLOW_SHITRANSCOEFF_3D
 
 C     allow code for simple ISOMIP thermodynamics
 #define ALLOW_ISOMIP_TD
@@ -23,6 +23,12 @@ C     allow code for simple ISOMIP thermodynamics
 C     allow friction velocity-dependent transfer coefficient
 C     following Holland and Jenkins, JPO, 1999
 #undef SHI_ALLOW_GAMMAFRICT
+
+C     allow (vertical) remeshing whenever ocean top thickness factor
+C     exceeds thresholds
+#undef ALLOW_SHELFICE_REMESHING
+C     and allow to print message to STDOUT when this happens
+#define SHELFICE_REMESH_PRINT
 
 #endif /* ALLOW_SHELFICE */
 #endif /* SHELFICE_OPTIONS_H */
