+++ verification/isomip/code_ad/packages.conf
@@ -3,6 +3,7 @@ gfd
 ggl90
 cd_code
 shelfice
+steep_icecavity
 diagnostics
 mnc
