+++ verification/isomip/input.stic/data.stic
@@ -1,0 +1,6 @@
+# ==========================================
+# | Parameters for STEEP_ICECAVITY package |
+# ==========================================
+ &STIC_PARM01
+ STICdepthFile = 'icetopo.exp1',
+ &
