+++ verification/isomip/input_ad.stic/data.pkg
@@ -1,0 +1,7 @@
+# Packages
+ &PACKAGES
+ useMNC=.TRUE.,
+ useShelfIce=.TRUE.,
+ useSTIC=.TRUE.,
+ useGrdchk=.TRUE.,
+ &
